backend/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.8.1
certifi==2025.8.3
# Pin to a stable version that has prebuilt Windows wheels for CPython 3.12
cffi==1.17.1
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.7.0
django-filter==24.3
django-minio-backend==3.8.1
django-minio-storage==0.5.8
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.11
Faker==35.2.0
inflection==0.5.1
iniconfig==2.1.0
minio==7.2.18
packaging==25.0
pluggy==1.6.0
psycopg==3.2.3
psycopg2-binary==2.9.10
pycparser==2.23
pycryptodome==3.23.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.5.0
requests>=2.31.0
django-guardian>=2.4.0
